Reorganizing code and imports
