+++ app/extensions/execute/native/native.py
@@ -11,5 +11,6 @@ class NativeExecutor(Extension):
         super().__init__([
             ('native.go', 'execute/native'),
             ('ls.go', 'execute/native/discovery'),
+            ('native_util.go', 'execute/native/util'),
         ])
         self.dependencies = []
